introduce custom function option for renderer plugins and moved puppeteer logic into plugin
