+++ packages/plugin-renderer-puppeteer/src/lib/browser.js
